Subst: Fix assertion failure
Substitution has a sanity check, verifying that different variables have
different names, because variables are compared by reference equality,
not structural equality. However, the traversal code will happily
recreate variables assuming that it doesn't make a difference -
assuming, that is, that only structural equality is used. A conflict
results.

To fix it, override reflectiveCopy to add a special case for unchanged
variables, to return the original value (even though arguably, this
optimization could be added to reflectiveCopy generically, to reduce
object churn).